+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Allow inference with multi-dimensional x when appropriate embedding is passed (#335)
 - Fixes a bug with clamp_and_warn not overriding num_atoms for SNRE and the warning message itself (#338)
 - Compatibility with Pyro 1.4.0 (#339)
-- Speed up posterior rejection sampling by introducing batch size (#340)
+- Speed up posterior rejection sampling by introducing batch size (#340, #343)
 - Allow vectorized evaluation of numpy potentials (#341)